Change Sound volume internally to float

+++ src/LoLTainer/LoLTainer/Interfaces/ISoundPlayer.cs
@@ -21,7 +21,7 @@ namespace LoLTainer.Interfaces
         /// <param name="playLengthInSec">seconds the file should be played</param>
         /// <param name="volume">volume standardized between 0 and 100. To not change the volume use -1.</param>
         /// <returns></returns>
-        Task PlaySound(int playerId, string fileName, TimeSpan? playLength, int volume = -1, PlayMode playMode = PlayMode.WaitPlaying);
+        Task PlaySound(int playerId, string fileName, TimeSpan? playLength, float volume = -1, PlayMode playMode = PlayMode.WaitPlaying);
         /// <summary>
         /// Stops the sound assosiated to a specific <paramref name="playerId"/>
         /// </summary>
